Refactor c++ utility functions
