+++ src/python/lib/CMakeLists.txt
@@ -4,7 +4,7 @@ addPythonModule(
     SOURCES
         z5py.cxx
         dataset.cxx
-        #groups.cxx
+        util.cxx
     LIBRRARIES
         ${COMPRESSION_LIBRARIES} 
         ${Boost_FILESYSTEM_LIBRARY}
